requirements.txt
yt-dlp
openai
torch
torchvision
torchaudio
python-docx
python-dotenv
openai-whisper
